feat: 가격 제안 플로우 통합
