+++ src/App.tsx
@@ -4,8 +4,7 @@ import Home from './pages/Home/Home';
 import BrandingFlow from './pages/BrandingFlow/BrandingFlow';
 import BrandResult from './pages/BrandResult/BrandResult';
 import MyPage from './pages/MyPage/MyPage';
-import PriceQuote from './pages/PriceQuote/PriceQuote';
-import PriceResult from './pages/PriceResult/PriceResult';
+import PriceQuoteFlow from './pages/PriceQuoteFlow/PriceQuoteFlow';
 import './App.css';
 
 const AppContainer = styled.div`
@@ -48,8 +47,7 @@ function App() {
             <Route path="/branding" element={<BrandingFlow />} />
             <Route path="/brand-result" element={<BrandResult />} />
             <Route path="/mypage" element={<MyPage />} />
-            <Route path="/price-quote" element={<PriceQuote />} />
-            <Route path="/price-result" element={<PriceResult />} />
+            <Route path="/price-quote" element={<PriceQuoteFlow />} />
             {/* 추가 페이지 라우트는 여기에 추가 */}
           </Routes>
         </ContentWrapper>
