+++ src/components/pricing/PriceQuoteStep/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PriceQuoteStep';
