--- src/pages/PriceQuote/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PriceQuote';
